requirements.txt
fastapi
uvicorn
psycopg2-binary>=2.9.11
pydantic>=2.12.5
pydantic-settings>=2.0.0
python-dotenv>=1.2.1
SQLAlchemy>=2.0.45
paho-mqtt>=2.1.0
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.20
